PIXIE: initial wrapper
Has to overload the use of the pose table. Would make
sense to refactor all of this so pose contains the
body pose for all of the variants, and then include
a model specific field for the extras (also for expose).

Nicely, this library works if all the installation
instructions are followed and everything is put in the
path, even when called from another directory.