Expose: render using cached SMPL parameters
This took much more work than expected and it appears the
Expose configuration parameters alter the interpretation
of the SMPL parameters when computing the meshes. This
might reflect a bug upstream, since really the body model
_should_ have a consistent interpretation. Nonetheless, the
joints computed from it should be robust and now the
visualization is correct without hacks to store all the
vertices in the databse.