PosePipeline: add OpenPose_LR option and up defaults
Playing with OpenPose_HR shows it works much better for geometric
consistency, so make this the default. Ultimately, we really want
to migrate entirely away from this class but that is still dependent
on a better face blurring.

Also add stock OpenPose settings (called for now) for comparison
but will probably consolidate these variations once testing is
complete.