[QR dev] Updating statistics tables
In the QR metrics table, adding an additional
field to pass more information to other tables
and prevent duplicate calculations. Also adding
check to modify window length if the number of
detections is less than the provided window_len
and to ensure there are at least 2 detections in
a video.