[devcontainer] Successfully creating devcontainer for PosePipeline
I added a command to the end of the dockerfile that runs
indefinitely. This got me past the issue of the container
not running when opening the devcontainer. I am not sure
if this is the best way to handle this, but I am looking
into it. Also need to clean up code.