[devcontainer] Able to connect to mysql database
I am now able to connect to the mysql database created
by the datajoint docker image. I added health check to
the docker compose file to make sure the database was
started before building the posepipeline container
but that may not be needed. I also install mysql-client
in the Dockerfile but that can be moved.