devcontainer: Updating Dockerfile and posepipeline path
For the Dockerfile, we decided to do devcontainer development
with the latest so I updated the base image to be Ubuntu 22.04 and
removed the pip downgrade. Additionally, I fixed the mmpretrain
installation issue by installing from source rather than mim, which
was recommended by openmmlab. Finally, I updated a path in __init__.py
in pose_pipeline.