devcontainer: Adding mmpose to devcontainer
In docker-compose for the devcontainer, allowing access to
more GPUs (this will change in the future). Adding mmpose
to the Dockerfile. Updating the mmdet wrapper, no functional
changes. Updating the mmpose wrapper to be compatible with the
new version of mmpose and adding COCO Wholebody RTMPose model.
Updating pipeline.py to use the RTMPose COCO Wholebody model.
Updating the getting started notebook to also run the mmpose
steps. Needs to be cleaned up.